My log and some observations:
Added Doxygen commenting to everything. Made the formatting and commenting more consistent, and cleaned up some random oddities.

(Note: The doxygen style is mixed, but doxygen does not care. I am using C++ " /// \ " style for the top headers, and the Javadoc/Chad's current style for everything else. We can switch either of these easily if desired.)

I did this in Visual Studio and cygwin threw no compilation errors, so I assume that there is no problem with Carriage Returns vs. Line Feeds. However, if you start getting junk, then this would be the problem, and is easy to fix.

Added a small ReadMe. This needs to change along with our project. The "true" documentation is Doxygen, however, I will be asking Hossain if this is sufficient, and if not, I will write up a manual documentation, which would not take long at all.

I believe Doxygen is sufficient, according to that one project I mentioned, and according to what others have told me.

I have encountered a few things that need to be mentioned, and changed a few things:

In main.cpp

I have altered the Main function to incorporate argc and argv. We need to parse and pass these parameters to the File reader. Simple, but a change I don't want to make before letting you guys know!



In admin.h and admin.cpp


-Changed the max error checking if-statements to a function call

-We should figure out the exact commenting style and/or spacing and bracket styles (I edited this in VS so we may have to resolve CR LF errors at some point. Not likely though.). Even Visual Studio seems to randomly switch between 2pt and 8pt tabbing for If statements.

-This might be my own ignorance, but should there not also be a check for BadSymbol in the admin cpp file? I mean, shouldn't bad symbol count as an error?
I have added such a check in the Illegal Character location ( || BADSYMBOL ). Feel free to remove it.

-I added in a "currentTokenType" that uses the getType once, that way we aren't using getType 4 or 5 times every single time we have a (valid) symbol to check.

-Since we are using an enum, we could change the if...else statements in the Admin class to a Switch statement. I have not done so as that's a slightly larger change and want to make sure you guys prefer it.

-The scanner bails out on an entire _line_ when an error is found. Perhaps we should go with what Hossain said instead, and go all the way to the next statement seperator (;) instead? Because we may have a wrong line and a correct line on the same actual line; an entire program could be 1 line!
This would eliminate that entire correctLine stuff, but we would need to automatically proceed to the char after the next terminal.

That is, reporting one error per line is fine, but disregarding any code after is sort of undesirable. We can fix this easily, but I don't want to fix it without your guys' presence in case I'm overlooking something.

-Currently, our error reporter seems to completely ignore $comment lines
in its error reporting. So when it says "Error at line x column y" it seems to disregard either comments or begin statements (further testing needed) or maybe the index is just wrong?



In files token.h and token.cpp

-I have changed the basic Enum layout, as I have lined up types of symbols with each other.
The counts in the comments were wrong (if the enum is sequential) as well.
Unless there is something I did not "get", in which case, feel free to revert the enum block.

-Do we need to have the Token types in the Protected area anymore? I think they can be private now since we do not subclass at all.

-I am sticking to a "comment what we see in the h, comment how it is done in the cpp" method of commenting, but it makes the files a lot larger and IMO  a bit harder to read. Is this OK?

-Changed the values in the definitions of the mutators to be self-descriptive instead of single chars. and to be consistent.



In symboltable.h and symboltable.cpp

-Shouldn't the get() function return a reference our point to the given symbol table instead of a copy?
Or does it already?
-In the cpp of makeEntry, can't we just return (index-1) at the end instead of declaring a new variable?


-The "getattribute" and "findattribute" functions seem redundant. We are returning a Lexeme string if the attribute is found
, otherwise we return an empty string. Why not make these into booleans? After all, we know what lexeme we will be using at the time we are searching for it. The "this is nothing, thus use the value nothing" should be decided at the caller level, not the callee level.
As well, the naming terminology is inconsistent.

-We frequently misues and mislabel Lexeme and Attribute. I tried to correct this a bit, but it's honestly confusing me as well. Overall, the Attribute Return functions are confusing to read and parse.



In scanner.h and scanner.cpp

-We include a boatload of libraries. Are these all actually needed?
For instance, unistd.h is wholly unneeded and unused.

-The scanner has a scan function that is not used anywhere.
Is this for the Parser in the future?

-We need to come up with a series of tests after correcting these issues. And kinds of tests - i.e., no white space tests, no newlines, broken code that bails out towards the end, errors before/after begin/end statements, etc.

-Uh, what's up with the reservedWords and tokenizedString functions and vector/array/variables?

-I changed the inRange function, so now it casts the source->size() to a signed integer; if we EVER got a negative size, then our filesystem would have to be horrendously corrupt or our CPU would be broken or something like that.


Overall I have kept this style: Class and Do-types are represented by a curly brace on the declare line, whereas if statements and functions are written as a curly brace after the declare line.


On the documentation end, I have forced doxygen to "Extract All" in case a given file has a syntax error.
I cannot seem to get it to display the code for CPP files. But it DOES grab the re-coded definitions and comments for the functions, you can see this in the greater Classes area.

We can add graphs at the end with HAVE_DOT = YES.